Fix a memory leak in `PostgreSQLConnection`
`PostgreSQLTableNameCache` holds a `PostgreSQLConnection` future. Once that future is fulfilled, it strongly holds a reference to `PostgreSQLConnection`. That connection in turn holds the table name cache, so we have a reference cycle. This causes the table name cache and its connection to never get released. This is relevant when testing, if the tests are set up such that each test creates a new `Database`.